move popover body to component
